# openapi.yaml
openapi: 3.0.3
info:
  title: ATC-Drone API
  version: "1.0.0"
  description: REST + WebSocket API for the ATC Drone Traffic Control system.
servers:
  - url: http://localhost:3000
...
paths:
...
  /v1/geofences/{id}:
    get:
      tags: [Geofences]
      summary: Get geofence
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Geofence
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Geofence"
    put:
      tags: [Geofences]
      summary: Update geofence
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeofenceRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Geofence"
    delete:
      tags: [Geofences]
      summary: Delete geofence
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: UUID
  schemas:
...
    Geofence:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        geofence_type:
          type: string
        polygon:
          type: array
          items:
            type: array
            items:
              type: number
        lower_altitude_m:
          type: number
        upper_altitude_m:
          type: number
        active:
          type: boolean
    GeofenceRequest:
      type: object
      properties:
        name:
          type: string
        geofence_type:
          type: string
        polygon:
          type: array
          items:
            type: array
            items:
              type: number
        lower_altitude_m:
          type: number
        upper_altitude_m:
          type: number
      required: [name, geofence_type, polygon]
...